modified:   INSTALL.ja 	modified:   Makefile 	modified:   src/Makefile 	modified:   src/config.h 	modified:   src/polyaness.c 	modified:   src/polyaness.h 	modified:   src/string.c 	modified:   src/string.h 	modified:   src/subset.c 	modified:   src/subset.h 	modified:   src/yasuna.c 	modified:   src/yasuna.h
