+++ INSTALL.ja
@@ -76,6 +76,18 @@ make に関する追加情報
   make 時のメッセージを抑制する為のオプション。
   未指定の場合の値は 0 であり、その他の値がセットされた場合は、抑制せずに出力される。
 
+  * BUFLEN
+  polyaness 内にてメモリの初期確保サイズ、および再確保の閾値を指定するオプション。
+  未指定の場合は 4096 であり、再確保の際は BUFLEN += BUFLEN となる。
+
+  * TH_LINES
+  p_read_file_char() 関数にて、メモリの初期確保サイズ(Y軸)、および再確保の閾値を指定するオプション。
+  未指定の場合は 1024 であり、再確保の際は TH_LINES += TH_LINES となる。
+
+  * TH_LENGTH
+  p_read_file_char() 関数にて、メモリの初期値確保サイズ(X軸)、および再確保の閾値を指定するオプション。
+  未指定の場合は 640 であり、再確保の際は TH_LENGTH += TH_LENGTH となる。
+
 * make のターゲット
 
   * all（未指定）
@@ -105,4 +117,20 @@ make に関する追加情報
   上記の install-* を実行する。
 
 
--- sasairc <sasairc@ssiserver.moe.hm>  Sun, 14 Feb 2016 18:14:38 +0900
+== 著作権
+
+            DO WHAT THE FUCK YOU WANT TO PUBLIC LICENSE
+                    Version 2, December 2004
+
+ Copyright (C) 2015 sasairc <sasairc@ssiserver.moe.hm>
+
+ Everyone is permitted to copy and distribute verbatim or modified
+ copies of this license document, and changing it is allowed as long
+ as the name is changed.
+
+            DO WHAT THE FUCK YOU WANT TO PUBLIC LICENSE
+   TERMS AND CONDITIONS FOR COPYING, DISTRIBUTION AND MODIFICATION
+
+  0. You just DO WHAT THE FUCK YOU WANT TO.
+
+-- sasairc <sasairc@ssiserver.moe.hm>  Sat, 20 May 2016 18:11:06 +0900
