+++ Makefile
@@ -2,20 +2,23 @@
 #    Makefile for yasuna
 #
 
-TARGET	= yasuna
-DICNME	= yasuna-quotes
+TARGET    = yasuna
+DICNME    = yasuna-quotes
 # DICNME = yasuna-quotes-lite
-PREFIX	:= /usr/local
-BINDIR	:= $(PREFIX)/bin
-DICDIR	:= $(PREFIX)/share/$(TARGET)
-MANDIR	:= $(PREFIX)/share/man/ja
-DOCDIR	:= $(PREFIX)/share/doc/$(TARGET)
-MAKE	:= make
-CC	:= cc
-RM	:= rm
-CFLAGS	:= -O2 -g -Wall
-LDFLAGS	:=
-CMDLINE	:= 0
+PREFIX    := /usr/local
+BINDIR    := $(PREFIX)/bin
+DICDIR    := $(PREFIX)/share/$(TARGET)
+MANDIR    := $(PREFIX)/share/man/ja
+DOCDIR    := $(PREFIX)/share/doc/$(TARGET)
+MAKE      := make
+CC        := cc
+RM        := rm
+CFLAGS    := -O2 -g -Wall
+LDFLAGS   :=
+CMDLINE   := 0
+BUFLEN    := 4096
+TH_LINES  := 1024
+TH_LENGTH := 640
 export
 
 all clean install-bin:
