+++ src/config.h
@@ -27,8 +27,14 @@
  * TH_LINES  : threshold of lines
  * TH_LENGTH : threshold of string (sizeof(char) * TH_LENGTH)
  */
+#ifndef TH_LINES
 #define TH_LINES        1024
+/* TH_LINES */
+#endif
+#ifndef TH_LENGTH
 #define TH_LENGTH       640
+/* TH_LENGTH */
+#endif
 
 #undef MONO             /* パスを組み込みたくないときに使うといいかもネ */
