+++ src/polyaness.h
@@ -45,6 +45,7 @@ typedef struct POLYANESS_T {
 extern int init_polyaness(FILE* fp, int offset, polyaness_t** polyaness);
 extern int parse_polyaness(FILE* fp, int offset, polyaness_t** polyaness);
 extern char* get_polyaness(const char* key, int record, polyaness_t** polyaness);
+extern void release_polyaness_cell(polyaness_cell** record);
 extern void release_polyaness(polyaness_t* polyaness);
 
 #ifdef  __cplusplus
