+++ src/subset.h
@@ -22,7 +22,6 @@ extern int open_dict_file(char* path, FILE** fp);
 extern int read_dict_file(FILE* fp, polyaness_t** pt);
 extern int parse_dict_file(FILE* fp, polyaness_t** pt);
 extern int select_by_speaker(char* speaker, polyaness_t** src, polyaness_t** dest);
-extern int strisdigit(char* str);
 extern int create_rand(int lines);
 extern void print_all_quotes(polyaness_t* pt, yasuna_t* yasuna);
